Controllers and views refactor
